# social-media-automation/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Media Automation",
    "description": "Niche-agnostic AI content automation \u2014 generate + post videos to IG, FB, YouTube, TikTok, LinkedIn.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/credentials": {
      "get": {
        "tags": [
          "credentials"
        ],
        "summary": "List Credentials",
        "operationId": "list_credentials_api_credentials_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "provider_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CredentialsRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "credentials"
        ],
        "summary": "Create Credentials",
        "operationId": "create_credentials_api_credentials_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/credentials/{cred_id}": {
      "patch": {
        "tags": [
          "credentials"
        ],
        "summary": "Update Credentials",
        "operationId": "update_credentials_api_credentials__cred_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cred_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cred Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "credentials"
        ],
        "summary": "Delete Credentials",
        "operationId": "delete_credentials_api_credentials__cred_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cred_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cred Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/credentials/{cred_id}/test": {
      "post": {
        "tags": [
          "credentials"
        ],
        "summary": "Test Credentials",
        "description": "Smoke-test the credentials against the actual provider.\n\nCurrently supports `llm` and `image` kinds. Others return a generic OK if the\npayload decrypts cleanly (we can't test e.g. social tokens without performing\na real post).",
        "operationId": "test_credentials_api_credentials__cred_id__test_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cred_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cred Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CredentialsCreate": {
        "properties": {
          "provider_kind": {
            "type": "string",
            "title": "Provider Kind",
            "description": "llm | image | voice | music | social"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name",
            "description": "openai | pexels | elevenlabs | ..."
          },
          "label": {
            "type": "string",
            "maxLength": 64,
            "title": "Label",
            "default": "default"
          },
          "payload": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Payload",
            "description": "Dict of secret fields, e.g. {'api_key': 'sk-...'} for OpenAI. Stored encrypted; never returned."
          }
        },
        "type": "object",
        "required": [
          "provider_kind",
          "provider_name",
          "payload"
        ],
        "title": "CredentialsCreate"
      },
      "CredentialsRead": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tenant_id": {
            "type": "integer",
            "title": "Tenant Id"
          },
          "provider_kind": {
            "type": "string",
            "title": "Provider Kind"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "secret_preview": {
            "type": "string",
            "title": "Secret Preview",
            "description": "Last 4 chars of api_key prefixed with '...' for visual confirmation"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "id",
          "tenant_id",
          "provider_kind",
          "provider_name",
          "label",
          "secret_preview"
        ],
        "title": "CredentialsRead"
      },
      "CredentialsTestResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "detail": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "message",
          "provider"
        ],
        "title": "CredentialsTestResult"
      },
      "CredentialsUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "title": "CredentialsUpdate",
        "description": "Rotate the secret payload OR change the label."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "title": "PageMeta"
      },
      "Page_CredentialsRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CredentialsRead"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[CredentialsRead]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}